+++ src/main/kotlin/data/CsvFoodRepositoryImp.kt
@@ -4,12 +4,12 @@ import logic.repository.FoodRepository
 import model.Food
 
 class CsvFoodRepositoryImp(
-    private val foodCsvParser: FoodCsvParser
+    private val foodParser: FoodParser
 ) : FoodRepository {
     private val foods = mutableListOf<Food>()
     override fun getFoods(): Result<List<Food>> {
         if (foods.isEmpty()) {
-            foodCsvParser.parseCsvFileToFoods().also { foods.addAll(it) }
+            foodParser.parse().also { foods.addAll(it) }
         }
         return Result.success(foods)
     }
