+++ src/main/kotlin/data/FoodParser.kt
@@ -1,0 +1,7 @@
+package data
+
+import model.Food
+
+interface FoodParser {
+    fun parse(): List<Food>
+}
